requirements.txt
# Core dependencies
torch>=2.0.0
transformers>=4.30.0
sentence-transformers>=2.2.0
faiss-cpu>=1.7.4
numpy>=1.24.0
pandas>=2.0.0

# LLM inference
llama-cpp-python>=0.1.78

# Web interface
streamlit>=1.28.0

# Data processing
beautifulsoup4>=4.12.0
lxml>=4.9.0
tiktoken>=0.5.0

# Utilities
python-dotenv>=1.0.0
tqdm>=4.65.0

# Kaggle API
kaggle>=1.5.15

